# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "FastAPI",
      "version": "0.1.0"
    },
    "paths": {
...
      "/requirement-states/": {
        "get": {
          "tags": [
            "requirement-states"
          ],
          "summary": "Get Requirement States",
          "operationId": "get_requirement_states_requirement_states__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/RequirementStateResponse"
                    },
                    "type": "array",
                    "title": "Response Get Requirement States Requirement States  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "requirement-states"
          ],
          "summary": "Create Requirement State",
          "operationId": "create_requirement_state_requirement_states__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementStateCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RequirementStateResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/requirement-states/{state_id}": {
        "get": {
          "tags": [
            "requirement-states"
          ],
          "summary": "Get Requirement State",
          "operationId": "get_requirement_state_requirement_states__state_id__get",
          "parameters": [
            {
              "name": "state_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "State Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RequirementStateResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "requirement-states"
          ],
          "summary": "Delete Requirement State",
          "operationId": "delete_requirement_state_requirement_states__state_id__delete",
          "parameters": [
            {
              "name": "state_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "State Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "requirement-states"
          ],
          "summary": "Update Requirement State",
          "operationId": "update_requirement_state_requirement_states__state_id__put",
          "parameters": [
            {
              "name": "state_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "State Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementStateUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RequirementStateResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/requirements/": {
        "get": {
          "tags": [
            "requirements"
          ],
          "summary": "Get Requirements",
          "operationId": "get_requirements_requirements__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/RequirementResponse"
                    },
                    "type": "array",
                    "title": "Response Get Requirements Requirements  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "requirements"
          ],
          "summary": "Create Requirement",
          "operationId": "create_requirement_requirements__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RequirementResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/requirements/{requirement_id}": {
        "get": {
          "tags": [
            "requirements"
          ],
          "summary": "Get Requirement",
          "operationId": "get_requirement_requirements__requirement_id__get",
          "parameters": [
            {
              "name": "requirement_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Requirement Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RequirementResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "requirements"
          ],
          "summary": "Delete Requirement",
          "operationId": "delete_requirement_requirements__requirement_id__delete",
          "parameters": [
            {
              "name": "requirement_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Requirement Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/reports/": {
        "get": {
          "tags": [
            "reports"
          ],
          "summary": "Get Reports",
          "operationId": "get_reports_reports__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/ReportResponse"
                    },
                    "type": "array",
                    "title": "Response Get Reports Reports  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "reports"
          ],
          "summary": "Create Report",
          "operationId": "create_report_reports__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ReportResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/reports/{report_id}": {
        "get": {
          "tags": [
            "reports"
          ],
          "summary": "Get Report",
          "operationId": "get_report_reports__report_id__get",
          "parameters": [
            {
              "name": "report_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Report Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ReportResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "reports"
          ],
          "summary": "Delete Report",
          "operationId": "delete_report_reports__report_id__delete",
          "parameters": [
            {
              "name": "report_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Report Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "reports"
          ],
          "summary": "Update Report",
          "description": "Update a report",
          "operationId": "update_report_reports__report_id__put",
          "parameters": [
            {
              "name": "report_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Report Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ReportResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ReportCreate": {
          "properties": {
            "title": {
              "type": "string",
              "title": "Title"
            },
            "description": {
              "type": "string",
              "title": "Description"
            },
            "duration": {
              "type": "string",
              "format": "duration",
              "title": "Duration"
            },
            "dead_time": {
              "type": "string",
              "format": "duration",
              "title": "Dead Time"
            },
            "dead_time_cause": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dead Time Cause"
            },
            "project_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            },
            "responsible_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Responsible Id"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            }
          },
          "type": "object",
          "required": [
            "title",
            "description",
            "duration",
            "dead_time"
          ],
          "title": "ReportCreate",
          "description": "Model for creating a report"
        },
        "ReportResponse": {
          "properties": {
            "title": {
              "type": "string",
              "title": "Title"
            },
            "description": {
              "type": "string",
              "title": "Description"
            },
            "duration": {
              "type": "string",
              "format": "duration",
              "title": "Duration"
            },
            "dead_time": {
              "type": "string",
              "format": "duration",
              "title": "Dead Time"
            },
            "dead_time_cause": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dead Time Cause"
            },
            "project_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            },
            "responsible_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Responsible Id"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            },
            "id": {
              "type": "integer",
              "title": "Id"
            }
          },
          "type": "object",
          "required": [
            "title",
            "description",
            "duration",
            "dead_time",
            "created_at",
            "updated_at",
            "id"
          ],
          "title": "ReportResponse",
          "description": "Model for report response"
        },
        "ReportUpdate": {
          "properties": {
            "title": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Title"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "duration": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "duration"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Duration"
            },
            "dead_time": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "duration"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dead Time"
            },
            "dead_time_cause": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dead Time Cause"
            },
            "project_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            },
            "responsible_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Responsible Id"
            }
          },
          "type": "object",
          "title": "ReportUpdate",
          "description": "Model for updating a report"
        },
        "RequirementCreate": {
          "properties": {
            "project_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            },
            "request_date": {
              "type": "string",
              "format": "date-time",
              "title": "Request Date"
            },
            "requested_by": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requested By"
            },
            "state_id": {
              "type": "integer",
              "title": "State Id"
            },
            "closing_date": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Closing Date"
            }
          },
          "type": "object",
          "required": [
            "state_id"
          ],
          "title": "RequirementCreate"
        },
        "RequirementResponse": {
          "properties": {
            "project_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            },
            "request_date": {
              "type": "string",
              "format": "date-time",
              "title": "Request Date"
            },
            "requested_by": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requested By"
            },
            "state_id": {
              "type": "integer",
              "title": "State Id"
            },
            "closing_date": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Closing Date"
            },
            "id": {
              "type": "integer",
              "title": "Id"
            }
          },
          "type": "object",
          "required": [
            "state_id",
            "id"
          ],
          "title": "RequirementResponse"
        },
        "RequirementStateCreate": {
          "properties": {
            "name": {
              "type": "string",
              "title": "Name"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "order": {
              "type": "integer",
              "title": "Order",
              "default": 0
            },
            "active": {
              "type": "boolean",
              "title": "Active",
              "default": true
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            }
          },
          "type": "object",
          "required": [
            "name"
          ],
          "title": "RequirementStateCreate"
        },
        "RequirementStateResponse": {
          "properties": {
            "name": {
              "type": "string",
              "title": "Name"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "order": {
              "type": "integer",
              "title": "Order",
              "default": 0
            },
            "active": {
              "type": "boolean",
              "title": "Active",
              "default": true
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            },
            "id": {
              "type": "integer",
              "title": "Id"
            }
          },
          "type": "object",
          "required": [
            "name",
            "id"
          ],
          "title": "RequirementStateResponse"
        },
        "RequirementStateUpdate": {
          "properties": {
            "name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "order": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order"
            },
            "active": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          "type": "object",
          "title": "RequirementStateUpdate"
        },
...
      }
    }
  }